- added integrity check for --force-new-root option
git-svn-id: https://svn.berlios.de/svnroot/repos/kiwi/kiwi-head@790 92bc7f66-5d17-0410-ac0d-e081219a1bb3

+++ modules/KIWIXML.pm
@@ -1638,6 +1638,12 @@ sub createTmpDirectory {
 			rmdir $root;
 			if ( -e $root && -d $root && $main::ForceNewRoot ) {
 				$kiwi -> info ("Removing old root directory '$root'");
+				if (-e $root."/base-system") {
+					$kiwi -> failed();
+					$kiwi -> info  ("Mount point /base-system exists");
+					$kiwi -> failed();
+					return undef;
+				}
 				qx (rm -R $root);
 				$kiwi -> done();
 			}
